separate parser to lexer and syntaxParser
