+++ cs/Markdown/Parser/MarkdownParser.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Markdown.Lexer;
+using Markdown.SyntaxParser;
 using Markdown.Tokens;
 
 namespace Markdown.Parser
@@ -8,15 +9,18 @@ namespace Markdown.Parser
     public class MarkdownParser : IParser
     {
         private readonly ILexer lexer;
+        private readonly ISyntaxParser syntaxParser;
 
-        public MarkdownParser(ILexer lexer)
+        public MarkdownParser(ILexer lexer, ISyntaxParser syntaxParser)
         {
             this.lexer = lexer;
+            this.syntaxParser = syntaxParser;
         }
 
         public IEnumerable<Token> Parse(string text)
         {
-            throw new NotImplementedException();
+            var lexedTokens = lexer.Lex(text);
+            return syntaxParser.Parse(lexedTokens);
         }
     }
 }
