+++ cs/Markdown/SyntaxParser/ISyntaxParser.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using Markdown.Tokens;
+
+namespace Markdown.SyntaxParser
+{
+    public interface ISyntaxParser
+    {
+        IEnumerable<Token> Parse(IEnumerable<Token> tokens);
+    }
+}
