+++ cs/Markdown/SyntaxParser/SyntaxParser.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using Markdown.Tokens;
+
+namespace Markdown.SyntaxParser
+{
+    public class SyntaxParser : ISyntaxParser
+    {
+        public IEnumerable<Token> Parse(IEnumerable<Token> tokens)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
